Refactor: Improve project structure and backend/frontend separation
This commit introduces several changes to enhance the project's structure and clarify the separation between backend, frontend, and shared code, following SvelteKit best practices.

Key changes include:

1.  **Client-Specific Code Isolation:**
    *   Created `src/lib/client/utils/` and `src/lib/client/types/` directories.
    *   Moved client-side utilities (`pdfGenerator.ts`, `toastStore.ts`, `transitionUtils.ts`) into `src/lib/client/utils/`.
    *   Moved client-specific types (`Toast`, Svelte transition types) into `src/lib/client/types/`.

2.  **Type Definition Reorganization:**
    *   Split the monolithic `src/lib/types.ts` into domain-specific type files within `src/lib/types/` (e.g., `user.types.ts`, `link.types.ts`).
    *   This improves modularity and makes it easier to locate type definitions.

3.  **Environment Variable Usage Verification:**
    *   Confirmed that server-side environment variables (like `DATA_DIR`) are accessed correctly via `process.env` within server modules and are not exposed to the client.

4.  **Script Review:**
    *   Verified that `scripts/db-tools.js` is a standalone Node.js utility and does not improperly bundle frontend code.

5.  **Documentation Update:**
    *   Updated the "Project Structure" section in `README.md` to accurately reflect these changes and provide a clearer overview of the codebase organization.

These changes contribute to a more maintainable and understandable codebase, making it easier for developers to navigate and contribute effectively.
