+++ src/lib/client/types/toast.types.ts
@@ -1,0 +1,6 @@
+export type Toast = {
+  id: string;
+  message: string;
+  type: "info" | "success" | "warning" | "error";
+  duration: number;
+};
