+++ src/lib/client/utils/pdfGenerator.ts
