+++ src/lib/types/link.types.ts
@@ -1,0 +1,7 @@
+export type Link = {
+  id: string;
+  userId: string;
+  url: string;
+  title: string;
+  createdAt: string;
+};
