+++ src/lib/types/session.types.ts
@@ -1,0 +1,6 @@
+export type Session = {
+  id: string;
+  userId: string;
+  createdAt: string;
+  expiresAt: string;
+};
