+++ src/lib/types/user.types.ts
@@ -1,0 +1,7 @@
+export type User = {
+  id: string;
+  passwordHash: string;
+  createdAt: string;
+  email: string;
+  name: string;
+};
